+++ S2/codigo/Disco.h
@@ -12,11 +12,11 @@ private:
 
 public:
 	Disco();
-	Disco(int valor, string tipo);
+	Disco(float valor, string tipo);
 	void aceptar(VisitanteEquipo visitante);
-	int getPrecio();
+	float getPrecio();
 	string getTipo();
-	void setPrecio(int precio);
+	void setPrecio(float precio);
 	void setTipo(string tipo);
 };
